+++ R/node_gaussian.r
@@ -2,13 +2,13 @@
 ## a node modeled using linear regression
 #' @export
 node_gaussian <- function(data, parents, formula=NULL, betas,
-                          intercept, error, var_corr=NULL) {
+                          intercept, error, var_corr=NULL, link="identity") {
 
   # if formula includes random effects, use node_lmer() instead
   if (!is.null(formula) & !is.null(var_corr)) {
     out <- node_lmer(data=data, formula=formula, betas=betas,
                      intercept=intercept, var_corr=var_corr, error=error,
-                     family="gaussian")
+                     family=stats::gaussian(link=link))
     return(out)
   }
 
@@ -24,7 +24,14 @@ node_gaussian <- function(data, parents, formula=NULL, betas,
   }
 
   out <- intercept +
-    rowSums(mapply("*", data, betas)) +
-    stats::rnorm(n=nrow(data), mean=0, sd=error)
+    rowSums(mapply("*", data, betas))
+
+  if (link=="log") {
+    out <- exp(out)
+  } else if (link=="inverse") {
+    out <- 1 / out
+  }
+
+  out <- out + stats::rnorm(n=nrow(data), mean=0, sd=error)
   return(out)
 }