+++ R/node_poisson.r
@@ -2,13 +2,13 @@
 ## a node modeled using poisson-regression
 #' @export
 node_poisson <- function(data, parents, formula=NULL, betas, intercept,
-                         var_corr=NULL) {
+                         var_corr=NULL, link="log") {
 
   # if formula includes random effects, use node_lmer() instead
   if (!is.null(formula) & !is.null(var_corr)) {
     out <- node_lmer(data=data, formula=formula, betas=betas,
                      intercept=intercept, var_corr=var_corr,
-                     family="poisson")
+                     family=poisson(link=link))
     return(out)
   }
 
@@ -23,8 +23,14 @@ node_poisson <- function(data, parents, formula=NULL, betas, intercept,
     data <- as.data.frame(data[, parents, with=FALSE])
   }
 
-  eff <- rowSums(mapply("*", data, betas))
-  mu <- exp(intercept + eff)
+  mu <- intercept + rowSums(mapply("*", data, betas))
+
+  if (link=="log") {
+    mu <- exp(mu)
+  } else if (link=="sqrt") {
+    mu <- mu^2
+  }
+
   out <- vapply(mu, FUN=stats::rpois, FUN.VALUE=numeric(1), n=1)
 
   return(out)